+++ src/main/java/com/Ultra_Nerd/CodeLyokoRemake15/world/generators/VolcanoGenerator.java
@@ -28,7 +28,7 @@ public class VolcanoGenerator extends NoiseChunkGenerator<VolcanoGenSettings> {
     @Override
     @Nonnull
     protected double[] getBiomeNoiseColumn(int noiseX, int noiseZ) {
-        return new double[]{(double)this.biomeProvider.func_222365_c(noiseX, noiseZ), 0.0D};
+        return new double[]{(double) this.biomeProvider.func_222365_c(noiseX, noiseZ), 0.0D};
     }
 
     @Override
@@ -38,14 +38,13 @@ public class VolcanoGenerator extends NoiseChunkGenerator<VolcanoGenSettings> {
 
     @Override
     protected double func_222551_g() {
-        return (double)((int)super.func_222551_g() / 2);
+        return (int) super.func_222551_g() / 2;
     }
 
     @Override
     protected double func_222553_h() {
         return 8.0D;
     }
-
 
 
     @Override
